--- GoReceitas/CollectionViewCells/FooterViewCell.swift
@@ -1,5 +1,0 @@
-import UIKit
-
-class FooterViewCell: UICollectionReusableView {
-    static let identifier = String(describing: FooterViewCell.self)
-}
