+++ GoReceitas/CollectionViewCells/Home/FooterViewCell.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+class FooterViewCell: UICollectionReusableView {
+    static let identifier = String(describing: FooterViewCell.self)
+}
