+++ GoReceitas/Model/AllTagsModel.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct AllTagsModel {
+    var image: String
+    var title: String
+    var subtitle: String
+}
